kevin_ai_bot-main/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.10.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
email-validator==2.3.0
python-multipart==0.0.22
PyPDF2==3.0.1
openai==1.99.9
google-auth==2.49.1
httpx==0.28.1
faster-whisper==1.1.1
